Currently working on the basedagent.py.
This is the core of the app's functionality.

Bascially, when the websocket connection is tarted, the files, messages, etc are loaded.

Upon each new message from the user, there is first a context filter agent, the context triage agent, which determines the relevant context to pass on to the based generator agent.

After the generator generates the response, if it is a .based or .based diff, the based code is validated and the diff as well (if applicable).

This is now sent to back to the websocket file to handle upodating the files and messages.

After this, all that is left for the backend (for now) is the handle file upload, handle file delete, handle restore version, and on disconnect handling.